- Created new project - Moved resources to new project, except the ones needed in ShareCad. - There's still some issues that need to be resolved in ShareCad.UI.
