+++ ShareCad.UI/BootlegResourceManager.cs
@@ -4,14 +4,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ShareCad
+namespace ShareCad.UI
 {
     internal static class BootlegResourceManager
     {
         public static class Icons
         {
-            public const string ConnectIcon = @"Images\connect_icon.png";
-            public const string NoConnectionIcon = @"Images\noconnect_icon.png";
+            public const string IpIcon = @"Images\ip_icon.png";
+            public const string ShareIcon = @"Images\share_icon.png";
+            public const string ShareIconNew = @"Images\share_icon_new.png";
         }
     }
 }
